+++ hubdle-gradle-plugin/test/kotlin/com/javiersc/hubdle/kotlin/multiplatform/features/KotlinMultiplatformFeaturesComposeTest.kt
@@ -1,16 +1,17 @@
 package com.javiersc.hubdle.kotlin.multiplatform.features
 
 import com.javiersc.gradle.testkit.test.extensions.GradleTest
+import io.kotest.matchers.nulls.shouldNotBeNull
+import io.kotest.matchers.shouldBe
 import kotlin.test.Test
+import org.gradle.testkit.runner.TaskOutcome
 
 internal class KotlinMultiplatformFeaturesComposeTest : GradleTest() {
 
     @Test
     fun compose() {
         gradleTestKitTest("kotlin/multiplatform/features/compose") {
-            println("------ output starts ------")
-            println(gradlew("build").output)
-            println("------ output ends ------")
+            gradlew("build").task(":build").shouldNotBeNull().outcome.shouldBe(TaskOutcome.SUCCESS)
         }
     }
 }